requirements.txt
garminconnect>=0.2.19
requests>=2.31.0
openai>=1.0.0
google-cloud-storage>=2.18.0
